refactor: split main into routes and handlers
